Convert.go and lint
