+++ workflow/common/util.go
@@ -24,9 +24,8 @@ import (
 	"k8s.io/client-go/tools/remotecommand"
 	"sigs.k8s.io/yaml"
 
-	"github.com/argoproj/argo/pkg/apis/workflow"
-
 	"github.com/argoproj/argo/errors"
+	"github.com/argoproj/argo/pkg/apis/workflow"
 	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	"github.com/argoproj/argo/util"
 )
@@ -679,25 +678,4 @@ func GetTemplateHolderString(tmplHolder wfv1.TemplateHolder) string {
 	} else {
 		return fmt.Sprintf("%T (%s)", tmplHolder, tmplName)
 	}
-}
-
-func ConvertCronWorkflowToWorkflow(cronWf *wfv1.CronWorkflow) (*wfv1.Workflow, error) {
-	newTypeMeta := metav1.TypeMeta{
-		Kind:       workflow.WorkflowKind,
-		APIVersion: cronWf.TypeMeta.APIVersion,
-	}
-
-	newObjectMeta := metav1.ObjectMeta{}
-	newObjectMeta.GenerateName = cronWf.Name + "-"
-
-	newObjectMeta.Labels = make(map[string]string)
-	newObjectMeta.Labels[LabelCronWorkflow] = cronWf.Name
-
-	wf := &wfv1.Workflow{
-		TypeMeta:   newTypeMeta,
-		ObjectMeta: newObjectMeta,
-		Spec:       cronWf.Spec.WorkflowSpec,
-	}
-	wf.SetOwnerReferences(append(wf.GetOwnerReferences(), *metav1.NewControllerRef(cronWf, wfv1.SchemeGroupVersion.WithKind(workflow.CronWorkflowKind))))
-	return wf, nil
 }
